requirements.txt
numpy
tensorflow
matplotlib
python-xlib
Pillow
pytesseract
pynput